修改app/console/commands/mycontroller.php

+++ app/Console/Commands/MyController.php
@@ -30,7 +30,7 @@ class MyController extends Command
      * Create a new command instance.
      * @return void
      */
-    public function __construct ()
+    public function __construct()
     {
         parent::__construct();
     }
@@ -39,31 +39,31 @@ class MyController extends Command
      * Execute the console command.
      * @return mixed
      */
-    public function handle ()
+    public function handle()
     {
-        $controllerName = $this->argument( 'name' );
+        $controllerName = $this->argument('name');
         $isAdmin = false;
-        $isMakeModel = $this->option( 'm' );
-        $isMakeRequest = $this->option( 'r' );
-        if ( strpos( strtolower( $controllerName ), 'admin' ) !== false ) {
+        $isMakeModel = $this->option('m');
+        $isMakeRequest = $this->option('r');
+        if (strpos(strtolower($controllerName), 'admin') !== false) {
             $this->isAdmin = true;
-            $controllerName = str_replace( 'Admin/', '', $controllerName );
+            $controllerName = str_replace('Admin/', '', $controllerName);
         }
 
-        $this->controllerName = ucfirst( $controllerName );
+        $this->controllerName = ucfirst($controllerName);
         //写入控制器内容
         $this->setControllerContent();
 
-        if ( $isAdmin && $isMakeModel ) {
+        if ($isMakeModel) {
             //写入模型
             $this->setModelContent();
         }
-        if ( $isAdmin && $isMakeRequest ) {
+        if ($isMakeRequest) {
             //写入Request
             $this->setRequestContent();
         }
         $echoMsg = 'controller ' . $this->controllerName . ' created success!';
-        $echoMsg .= "\r\n" . 'Router content is : Route::resource(\'' . strtolower( $this->controllerName ) . '\',\'' . $this->controllerName . 'Controller\')';
+        $echoMsg .= "\r\n" . 'Router content is : Route::resource(\'' . strtolower($this->controllerName) . '\',\'' . $this->controllerName . 'Controller\')';
         echo $echoMsg;
     }
 
@@ -72,7 +72,7 @@ class MyController extends Command
      *
      * @param $isAdmin
      */
-    protected function setControllerContent ()
+    protected function setControllerContent()
     {
         //后台管理控制器
         $controllerPathBack = app_path() . '/Http/Controllers/Admin/' . $this->controllerName . 'Controller.php';
@@ -80,17 +80,22 @@ class MyController extends Command
         $controllerPathFront = app_path() . '/Http/Controllers/' . $this->controllerName . 'Controller.php';
 
         //前台控制器内容
-        $contentFront = file_get_contents( dirname( __FILE__ ) . '/frontController' );
+        $contentFront = file_get_contents(dirname(__FILE__) . '/frontController');
         //后台控制器内容
-        $contentBack = file_get_contents( dirname( __FILE__ ) . '/backController' );
+        $contentBack = file_get_contents(dirname(__FILE__) . '/backController');
 
-        $contentFront = $this->replaceContent( $contentFront );
-        $contentBack = $this->replaceContent( $contentBack );
+        $contentFront = $this->replaceContent($contentFront);
+        $contentBack = $this->replaceContent($contentBack);
 
         //写入前台控制器
-        file_put_contents( $controllerPathFront, $contentFront );
+        if (!file_exists($controllerPathFront)) {
+            file_put_contents($controllerPathFront, $contentFront);
+
+        }
         //写入后台控制器
-        file_put_contents( $controllerPathBack, $contentBack );
+        if (!file_exists($controllerPathBack)) {
+            file_put_contents($controllerPathBack, $contentBack);
+        }
     }
 
     /**
@@ -98,27 +103,36 @@ class MyController extends Command
      *
      * @param $isAdmin
      */
-    protected function setModelContent ()
+    protected function setModelContent()
     {
         $modelCommonPath = app_path() . '/Model/Common/Common' . $this->controllerName . '.php';
         $modelBackPath = app_path() . '/Model/Admin/' . $this->controllerName . '.php';
         $modelFrontPath = app_path() . '/Model/Web/' . $this->controllerName . '.php';
 
 
-        //前台控制器内容
-        $contentFrontBack = file_get_contents( dirname( __FILE__ ) . '/model' );
+        //后台模型
+        $contentBack = file_get_contents(dirname(__FILE__) . '/backmodel');
+        //前台模型
+        $contentFront = file_get_contents(dirname(__FILE__) . '/frontmodel');
         //后台控制器内容
-        $contentCommon = file_get_contents( dirname( __FILE__ ) . '/commonModel' );
+        $contentCommon = file_get_contents(dirname(__FILE__) . '/commonModel');
 
-        $contentFrontBack = $this->replaceContent( $contentFrontBack );
-        $contentCommon = $this->replaceContent( $contentCommon );
+        $contentBack = $this->replaceContent($contentBack);
+        $contentFront = $this->replaceContent($contentFront);
+        $contentCommon = $this->replaceContent($contentCommon);
 
         //写入前台模型
-        file_put_contents( $modelFrontPath, $contentFrontBack );
+        if (!file_exists($modelFrontPath)) {
+            file_put_contents($modelFrontPath, $contentFront);
+        }
         //写入公共模型
-        file_put_contents( $modelCommonPath, $contentCommon );
+        if (!file_exists($modelCommonPath)) {
+            file_put_contents($modelCommonPath, $contentCommon);
+        }
         //写入后台模型
-        file_put_contents( $modelBackPath, $contentFrontBack );
+        if (!file_exists($modelBackPath)) {
+            file_put_contents($modelBackPath, $contentBack);
+        }
     }
 
     /**
@@ -126,13 +140,15 @@ class MyController extends Command
      *
      * @param $isAdmin
      */
-    protected function setRequestContent ()
+    protected function setRequestContent()
     {
-        $content = file_get_contents( dirname( __FILE__ ) . '/request' );
+        $content = file_get_contents(dirname(__FILE__) . '/request');
         $filePath = app_path() . '/Http/Requests/' . $this->controllerName . 'Request.php';
-        $content = $this->replaceContent( $content );
+        $content = $this->replaceContent($content);
 
-        file_put_contents( $filePath, $content );
+        if (!file_exists($filePath)) {
+            file_put_contents($filePath, $content);
+        }
     }
 
     /**
@@ -142,15 +158,15 @@ class MyController extends Command
      *
      * @return mixed
      */
-    protected function replaceContent ( $content )
+    protected function replaceContent($content)
     {
-        $content = str_replace( [
+        $content = str_replace([
             '{controller}',
             '{here}',
         ], [
             $this->controllerName,
-            iconv( 'gb2312', 'utf-8', $this->option( 'h' ) ),
-        ], $content );
+            iconv('gb2312', 'utf-8', $this->option('h')),
+        ], $content);
 
         return $content;
     }
